static initialization and port turn LED on

+++ ref_app/src/mcal/atmega4809/mcal_benchmark.h
@@ -17,7 +17,7 @@
   {
     namespace benchmark
     {
-      typedef mcal::port::port_pin<std::uint8_t,
+      typedef mcal::port::port_pin<std::uint16_t,
                                    std::uint8_t,
                                    mcal::reg::portd,
                                    UINT8_C(3)> benchmark_port_type;
